Added default constructors to components so they could be used as beans.

+++ uiframework/src/org/cougaar/lib/uiframework/ui/models/TestFunctionTableModel.java
@@ -57,7 +57,7 @@ public class TestFunctionTableModel extends DefaultTableModel
         }
         else
         {
-            return Integer.class;
+            return Double.class;
         }
     }
